fix admin code
