+++ app/Services/AdminUserService.php
@@ -4,7 +4,6 @@ namespace App\Services;
 
 use App\Models\User;
 use App\Models\UserProfile;
-use App\Models\Post;
 use App\Models\Upload;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -26,8 +25,6 @@ class AdminUserService {
     protected $currentUserDetails;
 
     protected $users;
-    
-    protected $posts;
 
     protected $userProfile;
 
@@ -38,9 +35,6 @@ class AdminUserService {
         $this->currentUserDetails = Auth::user();
         // User model object
         $this->users = new User();
-
-        // post model object
-        $this->posts = new Post();
 
         // upload model object
         $this->upload = new Upload();
@@ -92,33 +86,6 @@ class AdminUserService {
                                   ->orderBy('id','DESC')
                                   ->paginate(10);
         return $userArray;
-    }
-
-    /**
-     * [getPostTotal] we are getiing number of total posts
-     * @param  
-     * @param  
-     * @return dataCount
-     */
-    public function getPostTotal(){
-
-        $postArray =  $this->posts->whereNull('deleted_at')                               
-                                  ->orderBy('created_at','ASC')
-                                  ->count();
-        return $postArray;
-    }
-    /**
-     * [getPostListing] we are getiing all the post
-     * @param  
-     * @param  
-     * @return dataArray
-     */
-    public function getPostsListing(){
-
-        $postArray =  $this->posts->whereNull('deleted_at')                               
-                                  ->orderBy('created_at','ASC')
-                                  ->paginate(10);
-        return $postArray;
     }
 
 
@@ -195,36 +162,6 @@ class AdminUserService {
         }
     }
     
-    /**
-     * upload image into directory
-     * @param  object  $input
-     * @return object array
-     */
-    public function getPostImage($image){
-        
-        $destinationPath = 'storage/images/';
-        $timeDate = strtotime(Carbon::now()->toDateTimeString());
-        $ext = $image->getClientOriginalExtension();
-        // $imageNameWithExt = $requestImageName->getClientOriginalName(); 
-        $filename = $timeDate.'_'.rand().'.'.$ext;
-        $image->move($destinationPath, $filename);
-        return $filename;
-    }
-    
-    /**
-     * upload video into directory
-     * @param  object  $video
-     * @return object array
-     */
-    public function getPostVideo($video){
-        $destinationPath = 'storage/videos/';
-        $timeDate = strtotime(Carbon::now()->toDateTimeString());
-        $filenameWithExt= $video->getClientOriginalName();
-        $extension = $video->getClientOriginalExtension();
-        $fileNameToStore = $filenameWithExt. '_'.$timeDate.'.'.$extension;
-        $path = $video->move($destinationPath,$fileNameToStore);
-        return $fileNameToStore;
-    }
 
      /**
      * Delete profile image if data not stor in db
@@ -336,151 +273,5 @@ class AdminUserService {
     }
 
 
-     /**
-     * get postImageArray to store
-     * @param  object  $imageArray,$post_id
-     * @return object array
-     */
-    public function getPostImageArray($imageArray,$post_id){
-        $newImageArray=[];
-        $oldUploadJsonString=Upload::where('post_id',$post_id)->get('image');
-        // dd(json_decode($oldUploadJsonString, true)[0]['image']);
-        if(!empty($imageArray)){
-            foreach($imageArray as $image){
-                $newImageArray[] = $this->getPostImage($image);
-            }
-            if(!empty($oldUploadJsonString) || $oldUploadJsonString!=null){
-                $oldPostImageArray=explode(' ', json_decode($oldUploadJsonString, true)[0]['image']);
-                $imageArray=array_merge($oldPostImageArray,$newImageArray);
-                return $imageArray;
-            }
-            else{
-                return $newImageArray;
-            }
-        }
-        // dd(array_merge($newImageArray,json_decode($oldUploadJsonString, true)));
-        $finalArray= array_merge($newImageArray,json_decode($oldUploadJsonString, true));
-        // dd($finalArray);
-        return (!empty($finalArray)) ? $finalArray[0] : $finalArray;
-    }
-
-     /**
-     * get postVideoArray to store
-     * @param  object  $imageArray,$post_id
-     * @return object array
-     */
-    public function getPostVideoArray($videoArray,$post_id){
-        $newVideoArray=[];
-        $oldUploadJsonString=Upload::where('post_id',$post_id)->get('video');
-        if(!empty($videoArray)){
-            foreach($videoArray as $video){
-                $newVideoArray[] = $this->getPostVideo($video);
-            }
-            if($oldUploadJsonString!=null || $oldUploadJsonString!=[] ){
-                $oldPostVideoArray=explode(' ', json_decode($oldUploadJsonString, true)[0]['video']);
-                $videoArray=array_merge($oldPostVideoArray,$newVideoArray);
-                return $videoArray;
-            }
-            else{
-                return $newVideoArray;
-            }
-        }
-        $finalArray =array_merge($newVideoArray,json_decode($oldUploadJsonString, true));
-        // dd($finalArray);
-        return (!empty($finalArray)) ? $finalArray[0] : $finalArray;
-    }
-
-
-     /**
-     * [savePost] we are storing the post Details from admin section 
-     * @param  requestInput get all the requested input data
-     * @param  message return message based on the condition 
-     * @return dataArray with message
-     */
-    public function savePost($input,$imageArray,$videoArray,&$message=''){
-        try{
-            $this->posts->post_type = $input['post_type'];
-            $this->posts->user_id = $input['user_id'];
-            $this->posts->post_text = $input['post_text'];
-            $this->posts->country_id =$input['country_id'];
-            $this->posts->surf_start_date = $input['surf_date'];
-            $this->posts->wave_size = $input['wave_size'];
-            $this->posts->board_type = $input['board_type'];
-            $this->posts->state_id = $input['state_id'];
-            $this->posts->local_beach_break_id = $input['local_beach_break_id'];
-            $this->posts->surfer = $input['surfer'];
-            $this->posts->optional_info = implode(" ",$input['optional_info']);
-            $this->posts->created_at = Carbon::now();
-            $this->posts->updated_at = Carbon::now();
-            
-            if($this->posts->save()){
-                //for store media into upload table
-                $post_id=$this->posts->id;
-                $newImageArray = $this->getPostImageArray($imageArray,$post_id);
-                $newVideoArray = $this->getPostVideoArray($videoArray,$post_id);
-                $this->upload->post_id = $this->posts->id;
-                $this->upload->image = ($newImageArray!=[]) ? implode(" ",$newImageArray) : null;
-                $this->upload->video = ($newVideoArray!=[]) ? implode(" ",$newVideoArray) : null;
-                
-                if($this->upload->save()){
-                    $message = 'Post has been created successfully.!';
-                    return true;
-                }
-                    
-             }
-                
-            }
-        catch (\Exception $e){     
-            // throw ValidationException::withMessages([$e->getPrevious()->getMessage()]);
-            $message='"'.$e->getMessage().'"';
-            return $message;
-        }
-    }
-
-    /**
-     * [updatePost] we are updating the post Details from admin section 
-     * @param  requestInput get all the requested input data
-     * @param  message return message based on the condition 
-     * @return dataArray with message
-     */
-    public function updatePost($input,$imageArray,$videoArray,$id,&$message=''){
-        $posts=$this->posts->find($id);
-        try{
-            $posts->post_type = $input['post_type'];
-            $posts->user_id = $input['user_id'];
-            $posts->post_text = $input['post_text'];
-            $posts->country_id =$input['country_id'];
-            $posts->surf_start_date = $input['surf_date'];
-            $posts->wave_size = $input['wave_size'];
-            $posts->board_type = $input['board_type'];
-            $posts->state_id = $input['state_id'];
-            $posts->local_beach_break_id = $input['local_beach_break_id'];
-            $posts->surfer = $input['surfer'];
-            $posts->optional_info = implode(" ",$input['optional_info']);
-            $posts->created_at = Carbon::now();
-            $posts->updated_at = Carbon::now();
-            
-            ///for updating media into upload table
-                $newImageArray = $this->getPostImageArray($imageArray,$posts->id);
-                $newVideoArray = $this->getPostVideoArray($videoArray,$posts->id);
-                
-                    Upload::where('post_id', $posts->id)
-                    ->update([
-                        'image'=>($newImageArray!=[]) ? implode(' ', array_filter($newImageArray)) : null,
-                        'video'=>($newVideoArray!=[]) ? implode(' ', array_filter($newVideoArray)) : null,
-                        ]);
-            
-            
-            if($posts->save()){
-                $message = 'Post has been updated successfully.!';
-                    return $message;
-                
-            }
-        }
-        catch (\Exception $e){     
-            // throw ValidationException::withMessages([$e->getPrevious()->getMessage()]);
-            $message='"'.$e->getMessage().'"';
-            return $message;
-        }
-    }
+    
 }
