+++ app/Services/PostService.php
@@ -1,0 +1,240 @@
+<?php
+
+namespace App\Services;
+
+use App\Models\User;
+use App\Models\UserProfile;
+use App\Models\Post;
+use App\Models\Upload;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Laravel\Fortify\Contracts\CreatesNewUsers;
+use Illuminate\Validation\ValidationException;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+use Carbon\Carbon;
+use Storage;
+use DB;
+
+class PostService {
+    /**
+     * Create a new Service instance.
+     *
+     * @return void
+     */
+    
+    protected $posts;
+
+    protected $uploads;
+
+    public function __construct() {
+
+        // post model object
+        $this->posts = new Post();
+
+        // upload model object
+        $this->upload = new Upload();
+    }
+
+    /**
+     * [getPostTotal] we are getiing number of total posts
+     * @param  
+     * @param  
+     * @return dataCount
+     */
+    public function getPostTotal(){
+
+        $postArray =  $this->posts->whereNull('deleted_at')                               
+                                  ->orderBy('created_at','ASC')
+                                  ->count();
+        return $postArray;
+    }
+    /**
+     * [getPostListing] we are getiing all the post
+     * @param  
+     * @param  
+     * @return dataArray
+     */
+    public function getPostsListing(){
+
+        $postArray =  $this->posts->whereNull('deleted_at')                               
+                                  ->orderBy('created_at','ASC')
+                                  ->paginate(10);
+        return $postArray;
+    }
+
+
+    /**
+     * upload image into directory
+     * @param  object  $input
+     * @return object array
+     */
+    public function getPostImage($image){
+        
+        $destinationPath = 'storage/images/';
+        $timeDate = strtotime(Carbon::now()->toDateTimeString());
+        $ext = $image->getClientOriginalExtension();
+        // $imageNameWithExt = $requestImageName->getClientOriginalName(); 
+        $filename = $timeDate.'_'.rand().'.'.$ext;
+        $image->move($destinationPath, $filename);
+        return $filename;
+    }
+    
+    /**
+     * upload video into directory
+     * @param  object  $video
+     * @return object array
+     */
+    public function getPostVideo($video){
+        $destinationPath = 'storage/videos/';
+        $timeDate = strtotime(Carbon::now()->toDateTimeString());
+        $filenameWithExt= $video->getClientOriginalName();
+        $extension = $video->getClientOriginalExtension();
+        $fileNameToStore = $filenameWithExt. '_'.$timeDate.'.'.$extension;
+        $path = $video->move($destinationPath,$fileNameToStore);
+        return $fileNameToStore;
+    }
+
+
+     /**
+     * get postImageArray to store
+     * @param  object  $imageArray,$post_id
+     * @return object array
+     */
+    public function getPostImageArray($imageArray,$post_id){
+        $newImageArray=[];
+        $oldUploadedJsonString=Upload::where('post_id',$post_id)->get('image');
+        if($imageArray){
+            foreach($imageArray as $image){ 
+                $newImageArray[] = $this->getPostImage($image);
+            }
+        }
+            if(Upload::where('post_id',$post_id)->exists()){
+                if( $oldUploadedJsonString!=null || $oldUploadedJsonString!=[] || $oldUploadedJsonString!=""){
+                    $oldPostImageArray=explode(' ', json_decode($oldUploadedJsonString, true)[0]['image']);
+                    $imageArray=array_merge($oldPostImageArray,$newImageArray);
+                    return $imageArray;
+            }}
+            return array_merge($newImageArray,json_decode($oldUploadedJsonString, true));
+    }
+
+     /**
+     * get postVideoArray to store
+     * @param  object  $imageArray,$post_id
+     * @return object array
+     */
+    public function getPostVideoArray($videoArray,$post_id){
+        $newVideoArray=[];
+        $oldUploadedJsonString=Upload::where('post_id',$post_id)->get('video');
+        if($videoArray){
+            foreach($videoArray as $video){
+                $newVideoArray[] = $this->getPostVideo($video);
+            }
+        }
+        if(Upload::where('post_id',$post_id)->exists()){
+            if($oldUploadedJsonString!=null || $oldUploadedJsonString!=[] || $oldUploadedJsonString!=""){
+                $oldPostVideoArray=explode(' ', json_decode($oldUploadedJsonString, true)[0]['video']);
+                $videoArray=array_merge($oldPostVideoArray,$newVideoArray);
+                return $videoArray;
+            }
+        }
+        return array_merge($newVideoArray,json_decode($oldUploadedJsonString, true));
+    }
+
+
+     /**
+     * [savePost] we are storing the post Details from admin section 
+     * @param  requestInput get all the requested input data
+     * @param  message return message based on the condition 
+     * @return dataArray with message
+     */
+    public function savePost($input,$imageArray,$videoArray,&$message=''){
+        try{
+            $this->posts->post_type = $input['post_type'];
+            $this->posts->user_id = $input['user_id'];
+            $this->posts->post_text = $input['post_text'];
+            $this->posts->country_id =$input['country_id'];
+            $this->posts->surf_start_date = $input['surf_date'];
+            $this->posts->wave_size = $input['wave_size'];
+            $this->posts->board_type = $input['board_type'];
+            $this->posts->state_id = $input['state_id'];
+            $this->posts->local_beach_break_id = $input['local_beach_break_id'];
+            $this->posts->surfer = $input['surfer'];
+            $this->posts->optional_info = implode(" ",$input['optional_info']);
+            $this->posts->created_at = Carbon::now();
+            $this->posts->updated_at = Carbon::now();
+            
+            if($this->posts->save()){
+                //for store media into upload table
+                $post_id=$this->posts->id;
+                $newImageArray = $this->getPostImageArray($imageArray,$post_id);
+                $newVideoArray = $this->getPostVideoArray($videoArray,$post_id);
+                $this->upload->post_id = $this->posts->id;
+                $this->upload->image = ($newImageArray!=[]) ? implode(" ",$newImageArray) : null;
+                $this->upload->video = ($newVideoArray!=[]) ? implode(" ",$newVideoArray) : null;
+                $this->upload->save();
+                
+                if($this->upload->save()){
+                    $message = 'Post has been created successfully.!';
+                    return $message;
+                }
+                    
+             }
+                
+            }
+        catch (\Exception $e){     
+            // throw ValidationException::withMessages([$e->getPrevious()->getMessage()]);
+            $message='"'.$e->getMessage().'"';
+            return $message;
+        }
+    }
+
+    /**
+     * [updatePost] we are updating the post Details from admin section 
+     * @param  requestInput get all the requested input data
+     * @param  message return message based on the condition 
+     * @return dataArray with message
+     */
+    public function updatePost($input,$imageArray,$videoArray,$id,&$message=''){
+        
+        $posts=$this->posts->find($id);
+        try{
+            $posts->post_type = $input['post_type'];
+            $posts->user_id = $input['user_id'];
+            $posts->post_text = $input['post_text'];
+            $posts->country_id =$input['country_id'];
+            $posts->surf_start_date = $input['surf_date'];
+            $posts->wave_size = $input['wave_size'];
+            $posts->board_type = $input['board_type'];
+            $posts->state_id = $input['state_id'];
+            $posts->local_beach_break_id = $input['local_beach_break_id'];
+            $posts->surfer = $input['surfer'];
+            $posts->optional_info = implode(" ",$input['optional_info']);
+            $posts->created_at = Carbon::now();
+            $posts->updated_at = Carbon::now();
+            
+            ///for updating media into upload table
+                $newImageArray = $this->getPostImageArray($imageArray,$id);
+                $newVideoArray = $this->getPostVideoArray($videoArray,$id);
+                
+                    Upload::where('post_id', $posts->id)
+                    ->update([
+                        'image'=>($newImageArray!=[]) ? implode(' ', array_filter($newImageArray)) : null,
+                        'video'=>($newVideoArray!=[]) ? implode(' ', array_filter($newVideoArray)) : null,
+                        ]);
+            
+            
+            if($posts->save()){
+                $message = 'Post has been updated successfully.!';
+                    return $message;
+                
+            }
+        }
+        catch (\Exception $e){     
+            // throw ValidationException::withMessages([$e->getPrevious()->getMessage()]);
+            $message='"'.$e->getMessage().'"';
+            return $message;
+        }
+    }
+}
