+++ composer.json
@@ -16,8 +16,7 @@
         "laravel/jetstream": "^1.3",
         "laravel/sanctum": "^2.6",
         "laravel/tinker": "^2.0",
-        "livewire/livewire": "^2.0",
-        "pbmedia/laravel-ffmpeg": "^7.4"
+        "livewire/livewire": "^2.0"
     },
     "require-dev": {
         "facade/ignition": "^2.3.6",
