+++ public/css/loader.css
@@ -32,6 +32,8 @@ form .error {
     color: green;
 }
 
+
+
 .reset-img {
     opacity: 1;
     position: absolute;
