+++ routes/web.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\admin\AdminDashboard;
 use App\Http\Controllers\admin\AdminUserController;
 use App\Http\Controllers\WelcomeFeedController;
 use App\Http\Controllers\admin\AdminPageController;
+use App\Http\Controllers\user\UserPostController;
 use App\Http\Controllers\admin\PostController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\user\UserController;
@@ -45,7 +46,7 @@ Route::group(['middleware' => ['auth:sanctum', 'verified', 'userAuth']], functio
 
     Route::get('/dashboard', [DashboardController::class, 'dashboard'])->name('dashboard');  
      
-    Route::post('/create-post', [PostController::class, 'store'])->name('storeVedioImagePost');
+    Route::post('/create-post', [UserPostController::class, 'store'])->name('storeVideoImagePost');
 
     Route::get('/user/change-password', [UserController::class, 'showChangePassword'])->name('showPassword');
